module_files/CV/requirements.txt
fastapi
uvicorn
torch
pandas
numpy
python-multipart
Pillow
opencv-python

